+++ src/Colors/Cmyk/Color.php
@@ -59,7 +59,7 @@ class Color extends AbstractColor
         }
 
         try {
-            $color = InputHandler::withDecoders([
+            $color = InputHandler::usingDecoders([
                 Decoders\StringColorDecoder::class,
             ])->handle($input);
         } catch (NotSupportedException) {